+++ libc++/stdlib/strtol.cc
@@ -2,5 +2,5 @@
 
 int32_t strtol(const char *str, char **endptr, int base)
 {
-  return strtoll(str, endptr, base);
+  return static_cast<int32_t>(strtoll(str, endptr, base));
 }
